Verify view connections in remote slots matching strategy
