+++ runtime/browser/demo/recipes.manifest
@@ -61,4 +61,4 @@ recipe
 view PageProducts of [Product] in 'products.json'
 view ClairesWishlist of [Product] in 'wishlist.json'
 view APerson of Person in 'people.json'
-view ThePeople of Person in 'people.json'
+#view ThePeople of Person in 'people.json'
